refactor(frontend): move AuthTokenAcquiryInformation to a separate file
